admin updated

+++ DoctorBookingApp/admin-microservice/src/main/java/in/doctorbooking/ust/controller/AdminController.java
@@ -3,6 +3,7 @@ package in.doctorbooking.ust.controller;
 import in.doctorbooking.ust.domain.Doctor;
 import in.doctorbooking.ust.dto.DoctorDto;
 import in.doctorbooking.ust.dto.RequestDto;
+import in.doctorbooking.ust.exceptions.DoctorNotFoundException;
 import in.doctorbooking.ust.service.AdminService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -62,16 +63,70 @@ public class AdminController {
         return ResponseEntity.status(HttpStatus.OK).body(doctorDtoList);
     }
 
-    @GetMapping("/{doctorName}")
-    public ResponseEntity<DoctorDto> getDoctorByName(@PathVariable String doctorName) {
-        Optional<Doctor> newdoctor = Optional.of(adminService.findDoctorByName(doctorName));
-        return ResponseEntity.status(HttpStatus.OK).body(convertToDto(newdoctor.get()));
+    @GetMapping("doctorname/{doctorName}")
+    public ResponseEntity<List<DoctorDto>> getDoctorByName(@PathVariable String doctorName) {
+        final var doctorList = adminService.findByDoctorName(doctorName);
+        List<DoctorDto> doctorDtoList = doctorList.stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.status(HttpStatus.OK).body(doctorDtoList);
+
+    }
+    @GetMapping("doctorlocation/{doctorLocation}")
+    public ResponseEntity<List<DoctorDto>> getDoctorByLocation(@PathVariable String doctorLocation) {
+        final var doctorList = adminService.getDoctorByLocation(doctorLocation);
+        List<DoctorDto> doctorDtoList = doctorList.stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.status(HttpStatus.OK).body(doctorDtoList);
 
     }
 
-    @PutMapping("/{doctorName}")
-    public ResponseEntity<DoctorDto> updateDoctor(@PathVariable String doctorName, @Valid @RequestBody RequestDto doctorDto) {
-        Doctor existingDoctor = adminService.findDoctorByName(doctorName);
+    @GetMapping("/searchnamespec/{doctorName}/{doctorSpecialization}")
+    public ResponseEntity<List<DoctorDto>>searchDoctorByNameandSpec(@PathVariable String doctorName,@PathVariable String doctorSpecialization) {
+        final var doctorList = adminService.getDoctorByNameAndSpec(doctorName,doctorSpecialization);
+        List<DoctorDto> doctorDtoList = doctorList.stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.status(HttpStatus.OK).body(doctorDtoList);
+    }
+    @GetMapping("/searchlocspec/{doctorLocation}/{doctorSpecialization}")
+    public ResponseEntity<List<DoctorDto>>searchDoctorByLocandSpec(@PathVariable String doctorLocation,@PathVariable String doctorSpecialization) {
+        final var doctorList = adminService.getDoctorByLocAndSpec(doctorLocation,doctorSpecialization);
+        List<DoctorDto> doctorDtoList = doctorList.stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.status(HttpStatus.OK).body(doctorDtoList);
+    }
+
+    @GetMapping("/searchnameloc/{doctorName}/{doctorLocation}")
+    public ResponseEntity<List<DoctorDto>>searchDoctorByNameandLoc(@PathVariable String doctorName,@PathVariable String doctorLocation) {
+        final var doctorList = adminService.getDoctorByNameAndLoc(doctorName,doctorLocation);
+        List<DoctorDto> doctorDtoList = doctorList.stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.status(HttpStatus.OK).body(doctorDtoList);
+    }
+
+    @GetMapping("/searchnamelocspec/{doctorName}/{doctorLocation}/{doctorSpecialization}")
+    public ResponseEntity<List<DoctorDto>>searchDoctorByNameandLocandSpec(@PathVariable String doctorName,@PathVariable String doctorLocation,@PathVariable String doctorSpecialization) {
+        final var doctorList = adminService.getDoctorByNameAndLocAndSpec(doctorName,doctorLocation,doctorSpecialization);
+        List<DoctorDto> doctorDtoList = doctorList.stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.status(HttpStatus.OK).body(doctorDtoList);
+    }
+
+
+    @PutMapping("/id/{id}")
+    public ResponseEntity<DoctorDto> updateDoctorById(@PathVariable int id,@Valid @RequestBody RequestDto doctorDto) {
+        Doctor existingDoctor = adminService.findById(id);
         if (existingDoctor == null) {
             // Handle case where doctor is not found
             return ResponseEntity.notFound().build();
@@ -81,17 +136,19 @@ public class AdminController {
         return ResponseEntity.ok(convertToDto(updatedDoctor));
     }
 
-
-    @DeleteMapping("/delete/{doctorName}")
-    public ResponseEntity deleteDoctor(@PathVariable String doctorName) {
-        Optional<Doctor> newdoctor = Optional.of(adminService.findDoctorByName(doctorName));
-        adminService.remove(newdoctor.get());
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
+//    @DeleteMapping("/delete/{doctorName}")
+//    public ResponseEntity deleteDoctor(@PathVariable String doctorName) {
+//        Optional<Doctor> newdoctor = Optional.of(adminService.findDoctorByName(doctorName));
+//        adminService.remove(newdoctor.get());
+//        return ResponseEntity.status(HttpStatus.OK).build();
+//    }
 
     @DeleteMapping("/{id}")
     public ResponseEntity deleteDoctor(@PathVariable int id) {
         Optional<Doctor> newdoctor = Optional.of(adminService.findById(id));
+        if(newdoctor.isEmpty()){
+            throw new DoctorNotFoundException("No doctor to delete");
+        }
         adminService.remove(newdoctor.get());
         return ResponseEntity.status(HttpStatus.OK).build();
     }
